(chore) move ReviewPage out of subfolder, into components/Review
